# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SHEild AI 2.0 API Contract
  version: 0.1.0-mvp
  description: |
    SHEild AI 2.0 Phase 1 MVP Contract (Women-focused safety MVP).
    
    Architectural & Safety Guarantees:
    - Base URL: /api/v1
    - Auth: Bearer Supabase JWT (ES256/RS256 verified against JWKS)
    - Incident status and Notification/Alert status are strictly separated.
    - FCM push accepted != delivered != acknowledged. Provider acceptance only proves provider queued the notification.
    - Designed to be demographic-neutral so children, elderly, and other groups can be supported without breaking changes.
    - College minor project: zero paid service dependencies assumed.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development backend
paths:
...
  /journeys:
    get:
      summary: List visible journeys
      description: Owners see their journeys; accepted guardians see only journeys that selected them and remain unrevoked.
      operationId: listJourneys
      tags: [Safe Journey]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Journeys visible under RLS.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JourneyList' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      summary: Start a safe journey session
      description: |
        Enrolls a safe journey with scheduled ETA, grace period, and selected guardian notification.
      operationId: createJourney
      tags:
        - Safe Journey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJourneyRequest'
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Optional UUID deduplication key; if body client_event_id is also supplied it must match.
          schema: { type: string, format: uuid }
      responses:
        '201':
          description: Journey session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneyResponse'
        '200':
          description: Existing journey returned for a repeated client_event_id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JourneyResponse' }
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /journeys/{journey_id}:
    get:
      summary: Get journey status and details
      description: Retrieve safe journey status (in_progress, completed, missed_check_in, cancelled).
      operationId: getJourney
      tags:
        - Safe Journey
      security:
        - BearerAuth: []
      parameters:
        - name: journey_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Journey details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneyResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /journeys/{journey_id}/complete:
    post:
      summary: Complete an in-progress journey
      description: Owner-only conditional transition; retries after completion return the existing journey.
      operationId: completeJourney
      tags: [Safe Journey]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: journey_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Completed journey; repeating completion returns the same record.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JourneyResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { description: Journey is already in a different terminal state. }
  /journeys/{journey_id}/alerts:
    get:
      summary: List delivery status for missed-journey alerts
      description: Visible to the journey owner and selected accepted guardians under RLS.
      operationId: listJourneyAlerts
      tags: [Safe Journey]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: journey_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Journey alert queue progress.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/JourneyAlertItem' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /journeys/{journey_id}/cancel:
    post:
      summary: Cancel an in-progress journey
      operationId: cancelJourney
      tags: [Safe Journey]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: journey_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Cancelled journey; repeating cancellation returns the same record.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JourneyResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { description: Journey is already in a different terminal state. }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase user access token.
  responses:
    BadRequestError:
      description: Invalid request payload or parameter format.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid coordinate values provided."
              request_id: "req_01j8k9x"
    UnauthorizedError:
      description: Missing or invalid Supabase Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Token has expired or is invalid."
              request_id: "req_01j8k9y"
    ForbiddenError:
      description: Access denied. User is not the owner or an authorized guardian.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: "FORBIDDEN"
              message: "You are not authorized to view or acknowledge this incident."
              request_id: "req_01j8k9z"
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: "NOT_FOUND"
              message: "Requested incident or journey does not exist."
              request_id: "req_01j8k90"
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        detail:
          type: string
          description: Compatibility message retained for existing clients.
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "BAD_REQUEST"
            message:
              type: string
              example: "Detailed error description."
            request_id:
              type: string
              nullable: true
              example: "req_01j8k9x"
...
    CreateJourneyRequest:
      type: object
      required:
        - destination_label
        - eta_at
      properties:
        client_event_id:
          type: string
          format: uuid
          description: Optional client-generated key. A retry returns the existing journey.
        destination_label:
          type: string
          example: "Hostel / Home"
        destination_lat:
          type: number
          minimum: -90
          maximum: 90
          nullable: true
        destination_lon:
          type: number
          minimum: -180
          maximum: 180
          nullable: true
        eta_at:
          type: string
          format: date-time
          description: Estimated time of arrival in UTC.
        grace_seconds:
          type: integer
          minimum: 0
          maximum: 86400
          default: 900
          description: Time in seconds after ETA before missed check-in status triggers alerts (e.g. 900 = 15m).
        guardian_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Specific guardians to notify upon missed check-in.
      description: Destination coordinates must be supplied together. ETA must be in the future. Omitted guardian_ids selects all currently accepted guardians.
    JourneyResponse:
      type: object
      required:
        - id
        - owner_id
        - destination_label
        - eta_at
        - grace_seconds
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        destination_label:
          type: string
        destination_lat:
          type: number
          nullable: true
        destination_lon:
          type: number
          nullable: true
        guardian_ids:
          type: array
          items: { type: string, format: uuid }
        eta_at:
          type: string
          format: date-time
        grace_seconds:
          type: integer
        status:
          type: string
          enum:
            - in_progress
            - completed
            - missed_check_in
            - cancelled
          example: "in_progress"
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
...
    JourneyList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/JourneyResponse' }
    JourneyAlertItem:
      type: object
      required: [id, journey_id, guardian_id, priority, state, attempt_count]
      properties:
        id: { type: string, format: uuid }
        journey_id: { type: string, format: uuid }
        guardian_id: { type: string, format: uuid }
        priority: { type: integer, minimum: 1 }
        state: { type: string, enum: [queued, provider_accepted, provider_failed] }
        attempt_count: { type: integer, minimum: 0 }
        next_attempt_at: { type: string, format: date-time, nullable: true }